LayoutTests/fast/mediacapturefromelement/CanvasCaptureMediaStream-clone-track-expected.txt


FAIL check clone track after captureStream() assert_equals: expected true but got false
